\name{Medoids}
\alias{Medoids}
\docType{data}
\title{Simulated Cluster Medoids Data for 'EC_patient'
}
\description{
  This data set represents a potential data frame of cluster medoids' responses to selected asset variables, as generated by the function 'EconomicClusters'.
}
\usage{data(Medoids)}
\format{
  A data frame with 12 observations on 5 variables. To simplify the analysis for the sake of demonstration, this data frame was generated manually, as follows:
  \describe{
    \item{\code{V2}}{ 0}
    \item{\code{V5}}{ 1}
    \item{\code{V6}}{ 1}
    \item{\code{V9}}{ a categorical variable with levels 1 to 4}
    \item{\code{V11}}{ a categorical variable with levels 1 to 4}
  }
  
}
\details{
  The data frame 'Medoids' represents households within a simulated population selected as cluster medoids by the 'EconomicClusters' algorithm. The columns describe the medoids' responses to the 5 variables selected by the model. This data frame was generated manually.
}
\source{
  This data set was simulated by the package authors in order to demonstrate the functionality of the 'EconomicClusters' package.
}
\seealso{
\code{\link[EconomicClusters:EconomicClusters-package]{EconomicClusters}}
}
\examples{
#We previously defined an economic clustering model for our population using 'EconomicClusters'.
#We saved a data frame containing the responses to the economic questions 
#selected by the model for the cluster medoids as 'Medoids'.
#We have since collected trauma registry data for 250 patients 
#and asked them the 5 economic questions selected by the model.
#Data set 'Pts' contains the new patients' responses to these five questions.
#We now want to know which economic cluster each patient belongs in.

data(Pts)
data(Medoids)
Pt_clust<-EC_patient(Pts, Medoids)
#number of patients per cluster:
table(Pt_clust$group)

#We now have a vector of patient cluster membership ('Pt_clust$group')
}
\keyword{datasets}
