\name{Pop}
\alias{Pop}
\docType{data}
\title{Simulated DHS Population Assets Data for 'EC_patient'
}
\description{
  This data set was designed to represent a data frame of the DHS population responses to the asset variables selected by the 'EconomicClusters' algorithm, as generated by the function 'EconomicClusters'.
}
\usage{data(Pop)}
\format{
A data frame with 29 observations on 5 variables. To simplify the analysis for the sake of demonstration, this data frame was generated manually to match the observations in Pts who do not match cluster medoids:
\describe{
\item{\code{V2}}{ a binary variable with levels 0, 1}
\item{\code{V5}}{ a binary variable with levels 0, 1}
\item{\code{V6}}{ a binary variable with levels 0, 1}
\item{\code{V9}}{ a categorical variable with levels 1 to 4}
\item{\code{V11}}{ a categorical variable with levels 1 to 4}
}

}
\details{
The data frame 'Pop' represents DHS population responses to the asset variables selected by the 'EconomicClusters' algorithm. Note: A real DHS dataset has many more observations than this. This data frame was generated manually to match the observations in Pts who do not match cluster medoids.
}
\source{
  This data set was simulated by the package authors in order to demonstrate the functionality of the 'EconomicClusters' package.
}
\seealso{
  \code{\link[EconomicClusters:EconomicClusters-package]{EconomicClusters}}
}
\examples{
#We previously defined an economic clustering model for our population using 'EconomicClusters'.
#We saved a data frame containing the responses to the economic questions 
#selected by the model for the cluster medoids as Medoids.
#We saved a data frame consisting of the DHS population's responses to the economic questions
#selected by the model as Pop.
#We saved a vector of the cluster membership of each DHS observation as PopClusters.
#We have since collected trauma registry data for 250 patients 
#and asked them the 5 economic questions selected by the model.
#Data set Pts contains the new patients' responses to these five questions.
#We now want to know which economic cluster each patient belongs in.

data(Pts)
data(Medoids)
data(Pop)
data(PopClusters)
Pt_clust<-EC_patient(Pts, Medoids, Pop, PopClusters)
#number of patients per cluster:
table(Pt_clust$group)

#We now have a vector of patient cluster membership (Pt_clust$group).
}
\keyword{datasets}
